+++ lib/mega.dart
@@ -2,8 +2,7 @@ library mega;
 
 export 'package:mega/src/styles/mega_styles.dart';
 export 'package:mega/src/styles/mega_themes.dart';
-export 'package:mega/src/widgets/mega_bottom_navigation_bar_compact_item.dart';
-export 'package:mega/src/widgets/mega_bottom_navigation_bar_item.dart';
+export 'package:mega/src/widgets/mega_bottom_navigation_bar.dart';
 export 'package:mega/src/widgets/mega_button.dart';
 export 'package:mega/src/widgets/mega_card.dart';
 export 'package:mega/src/widgets/mega_divider.dart';
@@ -15,5 +14,7 @@ export 'package:mega/src/widgets/mega_linked_button.dart';
 export 'package:mega/src/widgets/mega_linked_buttons.dart';
 export 'package:mega/src/widgets/mega_linked_icon_button.dart';
 export 'package:mega/src/widgets/mega_linked_text_button.dart';
+export 'package:mega/src/widgets/mega_scaffold.dart';
+export 'package:mega/src/widgets/mega_separator.dart';
 export 'package:mega/src/widgets/mega_text_button.dart';
 export 'package:mega/src/widgets/mega_text_field.dart';