+++ lib/src/widgets/mega_divider.dart
@@ -3,27 +3,14 @@ import 'package:mega/src/styles/mega_styles.dart';
 import 'package:mega/src/utils/theme_picker.dart';
 
 class MegaHorizontalDivider extends StatelessWidget {
-  final double? height;
-  final double? indent;
-  final double? endIndent;
-
-  const MegaHorizontalDivider({
-    Key? key,
-    this.height,
-    this.indent,
-    this.endIndent,
-  }) : super(key: key);
+  const MegaHorizontalDivider({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return SizedBox(
-      height: height,
+      height: 1.0,
       child: Center(
         child: Container(
-          margin: EdgeInsetsDirectional.only(
-            start: indent ?? 0.0,
-            end: endIndent ?? 0.0,
-          ),
           decoration: BoxDecoration(
             color: ThemePicker.of(context).pick(
               light: MegaStyle.borderColorLight,
@@ -37,27 +24,14 @@ class MegaHorizontalDivider extends StatelessWidget {
 }
 
 class MegaVerticalDivider extends StatelessWidget {
-  final double? width;
-  final double? indent;
-  final double? endIndent;
-
-  const MegaVerticalDivider({
-    Key? key,
-    this.width,
-    this.indent,
-    this.endIndent,
-  }) : super(key: key);
+  const MegaVerticalDivider({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return SizedBox(
-      width: width,
+      width: 1.0,
       child: Center(
         child: Container(
-          margin: EdgeInsetsDirectional.only(
-            top: indent ?? 0.0,
-            bottom: endIndent ?? 0.0,
-          ),
           decoration: BoxDecoration(
             color: ThemePicker.of(context).pick(
               light: MegaStyle.borderColorLight,