+++ lib/src/widgets/mega_drawer_compact_item.dart
@@ -5,16 +5,16 @@ import 'package:mega/src/utils/theme_picker.dart';
 class MegaDrawerCompactItem extends StatefulWidget {
   final IconData icon;
   final Color accentColor;
-  final bool isActive;
+  final bool selected;
   final FocusNode? focusNode;
-  final bool? autofocus;
+  final bool autofocus;
   final VoidCallback? onPressed;
 
   const MegaDrawerCompactItem({
     Key? key,
     required this.icon,
     required this.accentColor,
-    this.isActive = false,
+    this.selected = false,
     this.focusNode,
     this.autofocus = false,
     required this.onPressed,
@@ -26,22 +26,20 @@ class MegaDrawerCompactItem extends StatefulWidget {
 
 class _MegaDrawerCompactItemState extends State<MegaDrawerCompactItem> {
   bool hovering = false;
-  bool disabled = true;
 
   @override
   Widget build(BuildContext context) {
-    if (widget.onPressed != null) disabled = false;
     return Padding(
       padding: MegaStyle.small,
       child: Stack(
         alignment: Alignment.centerLeft,
         children: [
           InkWell(
+            focusNode: widget.focusNode,
+            autofocus: widget.autofocus,
             onTap: widget.onPressed,
             onHover: (hover) => setState(() => hovering = hover),
             borderRadius: MegaStyle.borderRadius,
-            focusNode: widget.focusNode,
-            autofocus: widget.autofocus ?? false,
             child: AnimatedContainer(
               height: MegaStyle.buttonSize,
               width: MegaStyle.buttonSize,
@@ -49,44 +47,34 @@ class _MegaDrawerCompactItemState extends State<MegaDrawerCompactItem> {
               curve: MegaStyle.basicCurve,
               decoration: BoxDecoration(
                 borderRadius: MegaStyle.borderRadius,
-                color: disabled
+                color: widget.selected
                     ? ThemePicker.of(context).pick(
-                        light: MegaStyle.backgroundColorLight,
-                        dark: MegaStyle.backgroundColorDark,
+                        light: MegaStyle.hoverColorLight,
+                        dark: MegaStyle.hoverColorDark,
                       )
-                    : widget.isActive
+                    : hovering
                         ? ThemePicker.of(context).pick(
                             light: MegaStyle.hoverColorLight,
                             dark: MegaStyle.hoverColorDark,
                           )
-                        : hovering
-                            ? ThemePicker.of(context).pick(
-                                light: MegaStyle.hoverColorLight,
-                                dark: MegaStyle.hoverColorDark,
-                              )
-                            : ThemePicker.of(context).pick(
-                                light: MegaStyle.backgroundColorLight,
-                                dark: MegaStyle.backgroundColorDark,
-                              ),
+                        : ThemePicker.of(context).pick(
+                            light: MegaStyle.backgroundColorLight,
+                            dark: MegaStyle.backgroundColorDark,
+                          ),
               ),
               padding: MegaStyle.horizontal,
               child: Icon(
                 widget.icon,
                 size: MegaStyle.iconSize,
-                color: disabled
-                    ? ThemePicker.of(context).pick(
-                        light: MegaStyle.iconColorDisabledLight,
-                        dark: MegaStyle.iconColorDisabledDark,
-                      )
-                    : ThemePicker.of(context).pick(
-                        light: MegaStyle.iconColorLight,
-                        dark: MegaStyle.iconColorDark,
-                      ),
+                color: ThemePicker.of(context).pick(
+                  light: MegaStyle.iconColorLight,
+                  dark: MegaStyle.iconColorDark,
+                ),
               ),
             ),
           ),
           AnimatedContainer(
-            height: widget.isActive ? MegaStyle.iconSize : 0,
+            height: widget.selected ? MegaStyle.iconSize : 0,
             width: MegaStyle.halfPadding,
             duration: MegaStyle.basicDuration,
             curve: MegaStyle.basicCurve,