+++ lib/src/widgets/mega_linked_buttons.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:mega/src/styles/mega_styles.dart';
 import 'package:mega/src/utils/theme_picker.dart';
+import 'package:mega/src/widgets/mega_divider.dart';
 
 class MegaLinkedButtons extends StatelessWidget {
   final List<Widget> buttons;
@@ -9,6 +10,13 @@ class MegaLinkedButtons extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    List<Widget> addDividers(List<Widget> buttons) {
+      for (var i = 1; i < buttons.length; i = i + 2) {
+        buttons.insert(i, const MegaVerticalDivider());
+      }
+      return buttons;
+    }
+
     return Padding(
       padding: MegaStyle.hHorizontal,
       child: Container(
@@ -25,7 +33,7 @@ class MegaLinkedButtons extends StatelessWidget {
         ),
         child: ClipRRect(
           borderRadius: BorderRadius.circular(4.0),
-          child: Row(children: buttons),
+          child: Row(children: addDividers(buttons)),
         ),
       ),
     );