+++ lib/src/widgets/mega_linked_text_button.dart
@@ -4,15 +4,15 @@ import 'package:mega/src/utils/theme_picker.dart';
 
 class MegaLinkedTextButton extends StatefulWidget {
   final String text;
-  final bool isActive;
+  final bool selected;
   final FocusNode? focusNode;
-  final bool? autofocus;
+  final bool autofocus;
   final VoidCallback? onPressed;
 
   const MegaLinkedTextButton({
     Key? key,
     required this.text,
-    this.isActive = false,
+    this.selected = false,
     this.focusNode,
     this.autofocus = false,
     required this.onPressed,
@@ -30,11 +30,11 @@ class _MegaLinkedTextButtonState extends State<MegaLinkedTextButton> {
   Widget build(BuildContext context) {
     if (widget.onPressed != null) disabled = false;
     return InkWell(
+      focusNode: widget.focusNode,
+      autofocus: widget.autofocus,
       onTap: widget.onPressed,
       onHover: (hover) => setState(() => hovering = hover),
       borderRadius: MegaStyle.borderRadius,
-      focusNode: widget.focusNode,
-      autofocus: widget.autofocus ?? false,
       child: AnimatedContainer(
         height: MegaStyle.buttonSize,
         duration: MegaStyle.basicDuration,
@@ -45,7 +45,7 @@ class _MegaLinkedTextButtonState extends State<MegaLinkedTextButton> {
                   light: MegaStyle.backgroundColorDisabledLight,
                   dark: MegaStyle.backgroundColorDisabledDark,
                 )
-              : widget.isActive
+              : widget.selected
                   ? ThemePicker.of(context).pick(
                       light: MegaStyle.hoverColorLight,
                       dark: MegaStyle.hoverColorDark,